- moved the code into controllers
